% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_MOA.R
\name{DSD_MOA}
\alias{DSD_MOA}
\title{Base class for MOA-based data stream generators}
\usage{
DSD_MOA(...)
}
\arguments{
\item{...}{Further arguments (currently ignored).}
}
\value{
The abstract class cannot be instantiated and produces an error.
}
\description{
Abstract base class for MOA-based data stream generators. It inherits
directly from \link[stream:DSD]{stream::DSD}.
}
\examples{
\dontrun{
DSD_MOA()
}
}
\references{
Bifet A, Holmes G, Pfahringer B, Kranen P, Kremer H, Jansen T, and Seidl T
(2010). MOA: Massive Online Analysis, a Framework for Stream Classification
and Clustering. \emph{Journal of Machine Learning Research}, 11, 1601--1604.
}
\seealso{
Other DSD_MOA:
\code{\link[=DSD_RandomRBFGeneratorEvents]{DSD_RandomRBFGeneratorEvents()}}
}
\author{
Michael Hahsler
}
\concept{DSD_MOA}
