% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_RandomRBFGeneratorEvents.R
\name{DSD_RandomRBFGeneratorEvents}
\alias{DSD_RandomRBFGeneratorEvents}
\title{Random RBF Generator Events Data Stream Generator}
\usage{
DSD_RandomRBFGeneratorEvents(
  k = 3,
  d = 2,
  numClusterRange = 3L,
  kernelRadius = 0.07,
  kernelRadiusRange = 0,
  densityRange = 0,
  speed = 100L,
  speedRange = 0L,
  noiseLevel = 0.1,
  noiseInCluster = FALSE,
  eventFrequency = 30000L,
  eventMergeSplitOption = FALSE,
  eventDeleteCreate = FALSE,
  modelSeed = NULL,
  instanceSeed = NULL
)
}
\arguments{
\item{k}{Average number of centroids in the model.}

\item{d}{Number of dimensions in the generated stream.}

\item{numClusterRange}{Range for the number of clusters.}

\item{kernelRadius}{Average radius of the micro-clusters.}

\item{kernelRadiusRange}{Range of variation in micro-cluster radii.}

\item{densityRange}{Range of variation in cluster density.}

\item{speed}{Number of points between kernel movements.}

\item{speedRange}{Range of variation in kernel speed.}

\item{noiseLevel}{Proportion of noise points.}

\item{noiseInCluster}{If \code{TRUE}, allow noise points inside clusters.}

\item{eventFrequency}{Number of points between concept-drift events.}

\item{eventMergeSplitOption}{If \code{TRUE}, enable cluster merge and split events.}

\item{eventDeleteCreate}{If \code{TRUE}, enable cluster deletion and creation events.}

\item{modelSeed}{Random seed for the cluster model.}

\item{instanceSeed}{Random seed for generated instances.}
}
\value{
An object of class \code{DSD_RandomRBFGeneratorEvents} (subclass of
\link{DSD_MOA}, \link[stream:DSD]{stream::DSD}).
}
\description{
Generates random data using MOA's \code{RandomRBFGeneratorEvents} stream.
}
\details{
Only a subset of the parameters supported by the underlying MOA generator is
exposed. If \code{modelSeed} or \code{instanceSeed} is \code{NULL}, a seed is sampled from
R's random-number generator. Set these arguments explicitly to reproduce a
stream; call \code{set.seed()} to make the generated default seeds reproducible.

By default, the generator creates three clusters with concept drift. Cluster
locations move over time, and clusters may merge.
}
\examples{
stream <- DSD_RandomRBFGeneratorEvents()
get_points(stream, 10)

if (interactive()) {
animate_data(stream, n = 5000, horizon = 100, xlim = c(0, 1), ylim = c(0, 1))
}
}
\references{
Albert Bifet, Geoff Holmes, Bernhard
Pfahringer, Philipp Kranen, Hardy Kremer, Timm Jansen, Thomas Seidl.
MOA: Massive Online Analysis, a Framework for Stream
Classification and Clustering
\emph{Journal of Machine Learning Research (JMLR)}, 2010.
}
\seealso{
Other DSD_MOA:
\code{\link[=DSD_MOA]{DSD_MOA()}}
}
\author{
Michael Hahsler and John Forrest
}
\concept{DSD_MOA}
