% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSRegressor_MOA.R
\name{DSRegressor_MOA}
\alias{DSRegressor_MOA}
\alias{update.DSRegressor_MOA}
\alias{predict.DSRegressor_MOA}
\title{MOA-based stream regressor interface}
\usage{
DSRegressor_MOA(formula, RMOA_regressor)

\method{update}{DSRegressor_MOA}(object, dsd, n = 1, verbose = FALSE, block = 1000L, ...)

\method{predict}{DSRegressor_MOA}(object, newdata, type = "response", ...)
}
\arguments{
\item{formula}{Formula describing the regression problem.}

\item{RMOA_regressor}{A regressor object from \pkg{RMOA}.}

\item{object}{A \code{DSRegressor_MOA} object.}

\item{dsd}{A data stream object.}

\item{n}{Number of data points to read from the stream.}

\item{verbose}{If \code{TRUE}, report progress.}

\item{block}{Number of points processed per block.}

\item{...}{Further arguments passed to the underlying method.}

\item{newdata}{Data frame containing new observations.}

\item{type}{prediction type (see \code{\link[RMOA:predict.MOA_trainedmodel]{RMOA::predict.MOA_trainedmodel()}}).}
}
\value{
An object of class \code{DSRegressor_MOA}
}
\description{
Interface to MOA-based stream regression methods provided by the
\pkg{RMOA} package.
}
\details{
\code{DSRegressor_MOA} provides an interface to MOA-based stream regressors
through \pkg{RMOA}. See \link[RMOA:MOA_regressors]{RMOA::MOA_regressors} for available regressors.

Calls to \code{update()} train the current model incrementally.
}
\examples{
\dontrun{
library(streamMOA)
library(RMOA)

# create a data stream for the iris dataset
data <- iris[sample(nrow(iris)), ]
stream <- DSD_Memory(data)
stream

# define a stream regression model.
cl <- DSRegressor_MOA(
  Sepal.Length ~ Species + Sepal.Width + Petal.Length,
  RMOA::Perceptron()
  )

cl

# update the model with 100 points from the stream
update(cl, stream, 100)

# look at the RMOA model object
cl$RMOAObj

# make predictions for the next 50 points
newdata <- get_points(stream, n = 50)
pr <- predict(cl, newdata)
pr

plot(pr, newdata$Sepal.Length, xlim = c(0,10), ylim = c(0,10))
abline(a = 0, b = 1, col = "red")
}
}
\references{
Wijffels, J. (2014) Connect R with MOA to perform streaming
classifications. https://github.com/jwijffels/RMOA

Bifet A, Holmes G, Pfahringer B, Kranen P, Kremer H, Jansen T, Seidl T
(2010).  MOA: Massive Online Analysis, a Framework for Stream Classification
and Clustering. \emph{Journal of Machine Learning Research (JMLR)}.
}
\author{
Michael Hahsler
}
\concept{DSRegressor_MOA}
