% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_MOA.R
\name{DSD_MOA}
\alias{DSD_MOA}
\title{Base class for MOA-based Data Stream Generators}
\usage{
DSD_MOA(...)
}
\arguments{
\item{...}{further arguments.}
}
\value{
The abstract class cannot be instantiated and produces an error.
}
\description{
Abstract base class for MOA-based data stream generators
directly inherits from \link[stream:DSD]{stream::DSD}.
}
\examples{
DSD()
}
\references{
MOA: Massive Online Analysis, a Framework for Stream
Classification and Clustering Albert Bifet, Geoff Holmes, Bernhard
Pfahringer, Philipp Kranen, Hardy Kremer, Timm Jansen, Thomas Seidl.
Journal of Machine Learning Research (JMLR).
}
\seealso{
Other DSD_MOA: 
\code{\link{DSD_RandomRBFGeneratorEvents}()}
}
\author{
Michael Hahsler
}
\concept{DSD_MOA}
